pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyverto"
dynamic = ["version"]
description = "A python package version management tool."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["package version management", "version bump"]
authors = [
  { name = "phdenzel", email = "phdenzel@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "gitpython>=3.1.45",
    'tomli; python_version < "3.11"',
]

[project.scripts]
"pyverto" = "pyverto.__main__:main"

[project.urls]
Documentation = "https://github.com/phdenzel/pyverto#readme"
Issues = "https://github.com/phdenzel/pyverto/issues"
Source = "https://github.com/phdenzel/pyverto"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "ruff>=0.12.5",
]

[tool.hatch.version]
path = "src/pyverto/__about__.py"

[tool.hatch.build]
exclude = [
  ".github",
  "tests",
  "uv.lock",
  ".python-version",
  ".coverage",
]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/pyverto tests}"

# Linting, flaking, etc.
[tool.ruff]
line-length = 100
lint.select = ["E", "F", "D"]  # "D" includes pydocstyle rules
lint.ignore = ["D200", "D212", "D415"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

# Testing & coverage
[tool.coverage.run]
source_pkgs = ["pyverto", "tests"]
branch = true
parallel = true
omit = [
  "src/pyverto/__about__.py",
]

[tool.coverage.paths]
pyverto = ["src/pyverto", "*/pyverto/src/pyverto"]
tests = ["tests", "*/pyverto/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
